# apps/api/app/api/v1/billing.py
"""API v1 billing routes; no sibling billing API version exists yet."""

import json
from typing import Annotated, Any

from fastapi import APIRouter, Depends, Header, HTTPException, Path, Request, Response, status
from sqlmodel import Session

from app.api.deps import get_current_principal
from app.core.config import settings
from app.core.feature_flags import dev_routes_enabled, require_feature_enabled
from app.core.ratelimit import GENERAL_RATE_LIMIT, PAYMENTS_RATE_LIMIT, limiter
from app.core.security import Principal
from app.db.session import get_session
from app.integrations.payments import get_payments_client
from app.integrations.payments.mock import MockStripeClient, MockStripeError
from app.integrations.payments.webhooks import StripeWebhookSignatureError, construct_stripe_event
from app.models.domain import User
from app.schemas.billing import (
    BillingSubscriptionRead,
    CheckoutCreateRequest,
    CheckoutSessionRead,
    StripeWebhookRead,
)
from app.services.billing import StripeWebhookResult, apply_stripe_webhook_event
from app.services.llm_quota import normalized_email, resolve_user_for_principal

router = APIRouter(prefix="/billing", tags=["billing"])

PRO_MONTHLY_PRICE_CENTS = 1200


@router.get("/subscription")
@limiter.limit(GENERAL_RATE_LIMIT)
async def read_subscription(
    request: Request,
    response: Response,
    principal: Annotated[Principal, Depends(get_current_principal)],
    db_session: Annotated[Session, Depends(get_session)],
) -> BillingSubscriptionRead:
    del request, response
    require_feature_enabled("payments")
    user = resolve_user_for_principal(db_session, principal)
    return billing_subscription_to_read(user)


@router.post("/checkout")
@limiter.limit(PAYMENTS_RATE_LIMIT)
async def create_checkout(
    request: Request,
    response: Response,
    checkout_request: CheckoutCreateRequest,
    principal: Annotated[Principal, Depends(get_current_principal)],
    db_session: Annotated[Session, Depends(get_session)],
) -> CheckoutSessionRead:
    del request, response
    require_feature_enabled("payments")
    user = resolve_user_for_principal(db_session, principal)
    client = get_payments_client()
    session = await client.create_checkout_session(checkout_payload(checkout_request, user=user))
    return checkout_session_to_read(session, provider=client.provider)


@router.get("/checkout/{session_id}")
@limiter.limit(PAYMENTS_RATE_LIMIT)
async def retrieve_checkout(
    request: Request,
    response: Response,
    session_id: Annotated[str, Path(min_length=8)],
    principal: Annotated[Principal, Depends(get_current_principal)],
    db_session: Annotated[Session, Depends(get_session)],
) -> CheckoutSessionRead:
    del request, response
    require_feature_enabled("payments")
    user = resolve_user_for_principal(db_session, principal)
    client = get_payments_client()
    try:
        session = await client.retrieve_checkout_session(session_id)
    except MockStripeError as exc:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
    require_checkout_owner(session, user)
    return checkout_session_to_read(session, provider=client.provider)


@router.post("/checkout/{session_id}/complete")
@limiter.limit(PAYMENTS_RATE_LIMIT)
async def complete_mock_checkout(
    request: Request,
    response: Response,
    session_id: Annotated[str, Path(min_length=8)],
    principal: Annotated[Principal, Depends(get_current_principal)],
    db_session: Annotated[Session, Depends(get_session)],
) -> CheckoutSessionRead:
    del request, response
    require_feature_enabled("payments")
    user = resolve_user_for_principal(db_session, principal)
    if not dev_routes_enabled() or settings.stripe_secret_key:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Not found")
    try:
        mock_client = MockStripeClient()
        session = await mock_client.retrieve_checkout_session(session_id)
        require_checkout_owner(session, user)
        result = await mock_client.complete_checkout_session(session_id)
    except MockStripeError as exc:
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc)) from exc
    checkout_session = result["session"]
    process_mock_checkout_webhook(db_session, result["event"])
    return checkout_session_to_read(checkout_session, provider="mock")


@router.post("/webhooks/stripe")
async def stripe_webhook(
    request: Request,
    db_session: Annotated[Session, Depends(get_session)],
    stripe_signature: Annotated[str | None, Header(alias="Stripe-Signature")] = None,
) -> StripeWebhookRead:
    require_feature_enabled("payments")
    if not stripe_signature:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Missing Stripe-Signature header.",
        )
    try:
        payload = (await request.body()).decode("utf-8")
    except UnicodeDecodeError as exc:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Stripe webhook payload must be UTF-8 encoded.",
        ) from exc
    result = handle_stripe_webhook_payload(db_session, payload, stripe_signature)
    return stripe_webhook_to_read(result)


def process_mock_checkout_webhook(
    db_session: Session,
    event: dict[str, object],
) -> StripeWebhookResult:
    payload = stripe_event_payload(event)
    signature = MockStripeClient().sign_webhook_payload(payload)
    return handle_stripe_webhook_payload(db_session, payload, signature)


def handle_stripe_webhook_payload(
    db_session: Session,
    payload: str,
    stripe_signature: str,
) -> StripeWebhookResult:
    try:
        event = construct_stripe_event(payload, stripe_signature)
    except StripeWebhookSignatureError as exc:
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc)) from exc
    return apply_stripe_webhook_event(db_session, event)


def stripe_event_payload(event: dict[str, object]) -> str:
    return json.dumps(event, separators=(",", ":"), sort_keys=True)


def checkout_payload(request: CheckoutCreateRequest, *, user: User) -> dict[str, Any]:
    return {
        "mode": "subscription",
        "success_url": f"{settings.app_url}/en/billing/success?session_id={{CHECKOUT_SESSION_ID}}",
        "cancel_url": f"{settings.app_url}/en/billing",
        "customer_email": user.email,
        "client_reference_id": user.id,
        "metadata": {"plan": request.plan, "user_id": user.id},
        "line_items": [
            {
                "quantity": 1,
                "price_data": {
                    "currency": "usd",
                    "unit_amount": PRO_MONTHLY_PRICE_CENTS,
                    "recurring": {"interval": "month"},
                    "product_data": {"name": "Prompteer Pro"},
                },
            }
        ],
    }


def require_checkout_owner(session: dict[str, Any], user: User) -> None:
    customer_email = session.get("customer_email")
    if isinstance(customer_email, str) and normalized_email(customer_email) == user.email:
        return
    client_reference_id = session.get("client_reference_id")
    if isinstance(client_reference_id, str) and client_reference_id == user.id:
        return
    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No such checkout.session.")


def checkout_session_to_read(session: dict[str, Any], *, provider: str) -> CheckoutSessionRead:
    return CheckoutSessionRead(
        id=str(session["id"]),
        mode=str(session["mode"]),
        status=str(session["status"]),
        payment_status=str(session["payment_status"]),
        amount_total=session["amount_total"] if isinstance(session["amount_total"], int) else None,
        currency=session["currency"] if isinstance(session["currency"], str) else None,
        url=session["url"] if isinstance(session["url"], str) else None,
        customer_email=(
            session["customer_email"] if isinstance(session["customer_email"], str) else None
        ),
        provider=provider,
    )


def billing_subscription_to_read(user: User) -> BillingSubscriptionRead:
    return BillingSubscriptionRead(
        plan=user.plan,
        status="active" if user.plan == "paid" else "inactive",
        customer_email=user.email,
        provider="stripe" if settings.stripe_secret_key else "mock",
    )


def stripe_webhook_to_read(result: StripeWebhookResult) -> StripeWebhookRead:
    return StripeWebhookRead(
        received=True,
        event_id=result.event_id,
        event_type=result.event_type,
        processed=result.processed,
        customer_email=result.customer_email,
        user_id=result.user_id,
    )